# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/alerts": {
      "post": {
        "operationId": "AlertsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alert created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          },
          "409": {
            "description": "Alert with this threshold already exists for this item"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new alert",
        "tags": [
          "Alerts"
        ]
      },
      "get": {
        "operationId": "AlertsController_findAll",
        "parameters": [
          {
            "name": "itemId",
            "required": false,
            "in": "query",
            "description": "Filter alerts by item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get alerts",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/alerts/statistics/monthly": {
      "get": {
        "operationId": "AlertsController_getMonthlyStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Monthly alerts statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "month": {
                            "type": "string",
                            "example": "Jan"
                          },
                          "year": {
                            "type": "number",
                            "example": 2025
                          },
                          "count": {
                            "type": "number",
                            "example": 5
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 45
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get monthly alerts statistics",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/alerts/{id}": {
      "get": {
        "operationId": "AlertsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Alert ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Alert not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get alert by ID",
        "tags": [
          "Alerts"
        ]
      },
      "patch": {
        "operationId": "AlertsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Alert ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alert updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Alert not found"
          },
          "409": {
            "description": "Alert with this threshold already exists for this item"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update alert",
        "tags": [
          "Alerts"
        ]
      },
      "delete": {
        "operationId": "AlertsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Alert ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alert deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Alert not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/alerts/check": {
      "post": {
        "operationId": "AlertsController_checkAlerts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Alerts checked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "checkedAlerts": {
                      "type": "number"
                    },
                    "triggeredAlerts": {
                      "type": "number"
                    },
                    "sentAlerts": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to check alerts"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check all alerts and send notifications if needed",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/alerts/test-email": {
      "post": {
        "description": "Sends a test email to the configured recipient to verify that email functionality is working properly",
        "operationId": "AlertsController_testEmail",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "Email address to send test email to",
            "schema": {
              "example": "test@example.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "messageId": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email service not configured"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to send test email"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send test email to verify email configuration",
        "tags": [
          "Alerts"
        ]
      }
    },
...
  },
  "info": {
    "title": "ShelfSpot API",
    "description": "API for ShelfSpot application with JWT Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAlertDto": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number",
            "example": 1,
            "description": "Item ID for the alert"
          },
          "threshold": {
            "type": "number",
            "example": 5,
            "description": "Threshold quantity that triggers the alert",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "example": "Low stock alert",
            "description": "Alert name/description"
          }
        },
        "required": [
          "itemId",
          "threshold"
        ]
      },
      "UpdateAlertDto": {
        "type": "object",
        "properties": {
          "threshold": {
            "type": "number",
            "example": 5,
            "description": "Threshold quantity that triggers the alert",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "example": "Low stock alert",
            "description": "Alert name/description"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the alert is active"
          }
        }
      },
...
    }
  }
}